+++ src/main/java/org/fogbowcloud/saps/engine/core/scheduler/arrebol/Arrebol.java
@@ -1,7 +1,7 @@
 package org.fogbowcloud.saps.engine.core.scheduler.arrebol;
 
 import java.util.List;
-import org.fogbowcloud.saps.engine.core.dto.JobResponseDTO;
+import org.fogbowcloud.saps.engine.core.scheduler.executor.arrebol.dtos.JobResponseDTO;
 import org.fogbowcloud.saps.engine.core.model.SapsImage;
 import org.fogbowcloud.saps.engine.core.model.SapsJob;
 import org.fogbowcloud.saps.engine.core.scheduler.arrebol.exceptions.GetCountsSlotsException;
