+++ src/main/java/org/fogbowcloud/saps/engine/core/scheduler/arrebol/retry/GetJobByIdRetry.java
@@ -1,6 +1,6 @@
 package org.fogbowcloud.saps.engine.core.scheduler.arrebol.retry;
 
-import org.fogbowcloud.saps.engine.core.dto.JobResponseDTO;
+import org.fogbowcloud.saps.engine.core.scheduler.executor.arrebol.dtos.JobResponseDTO;
 import org.fogbowcloud.saps.engine.core.scheduler.arrebol.Arrebol;
 import org.fogbowcloud.saps.engine.core.scheduler.arrebol.exceptions.GetJobException;
